Ref: Created card_bottom_details and moved file inside
