+++ app/src/components/market/sections/common_sections/card_bottom_details/bar_diagram_probabilities/index.tsx
@@ -1,7 +1,7 @@
 import React, { DOMAttributes } from 'react'
 import styled from 'styled-components'
 
-import { getOutcomeColor } from '../../../../theme/utils'
+import { getOutcomeColor } from '../../../../../../theme/utils'
 
 const BarDiagramWrapper = styled.div`
   display: flex;
