+++ app/src/components/market/sections/market_buy/scalar_market_buy.tsx
@@ -18,13 +18,13 @@ import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { CurrencySelector } from '../../common/currency_selector'
-import { GridTransactionDetails } from '../../common/grid_transaction_details'
-import { MarketScale } from '../../common/market_scale'
 import { TransactionDetailsCard } from '../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../common/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
 import { WarningMessage } from '../../common/warning_message'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
+import { MarketScale } from '../common_sections/card_bottom_details/market_scale'
 
 const StyledButtonContainer = styled(ButtonContainer)`
   justify-content: space-between;
