+++ app/src/components/market/sections/market_create/steps/items/funding_and_fee_step.tsx
@@ -41,8 +41,6 @@ import {
 } from '../../../../common/common_styled'
 import { CreateCard } from '../../../../common/create_card'
 import { CurrencySelector } from '../../../../common/currency_selector'
-import { GridTransactionDetails } from '../../../../common/grid_transaction_details'
-import { MarketScale } from '../../../../common/market_scale'
 import { TradingFeeSelector } from '../../../../common/trading_fee_selector'
 import { TransactionDetailsCard } from '../../../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../../../common/transaction_details_line'
@@ -51,6 +49,8 @@ import { VerifiedRow } from '../../../../common/verified_row'
 import { WarningMessage } from '../../../../common/warning_message'
 import { SetAllowance } from '../../../common_sections/allowance/set_allowance'
 import { DisplayArbitrator } from '../../../common_sections/arbitrators/display_arbitrator'
+import { GridTransactionDetails } from '../../../common_sections/card_bottom_details/grid_transaction_details'
+import { MarketScale } from '../../../common_sections/card_bottom_details/market_scale'
 import { Outcome } from '../outcomes'
 
 const CreateCardTop = styled(CreateCard)`
