+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -13,13 +13,13 @@ import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { GenericError } from '../../common/common_styled'
-import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { OutcomeTable } from '../../common/outcome_table'
 import { TokenBalance } from '../../common/token_balance'
 import { TransactionDetailsCard } from '../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../common/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
 import { WarningMessage } from '../../common/warning_message'
+import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 
 const StyledButtonContainer = styled(ButtonContainer)`
   justify-content: space-between;
